@@ -2,10 +2,7 @@
   "name": "flight_plan",
   "private": true,
   "dependencies": {
-    "@babel/core": "^7.12.10",
-    "@babel/preset-react": "^7.12.10",
     "@githubprimer/octicons-react": "^8.1.0",
-    "@rails/webpacker": "^6.0.0-beta.5",
     "@reach/dialog": "^0.13.2",
     "@reach/menu-button": "^0.13.2",
     "@reach/portal": "^0.13.2",
@@ -16,14 +13,10 @@
     "actioncable": "5.2.4-5",
     "autoprefixer": "^10.2.1",
     "classnames": "^2.2.6",
-    "css-loader": "^5.0.1",
-    "css-minimizer-webpack-plugin": "^1.2.0",
     "downshift": "^6.1.0",
     "font-color-contrast": "^1.0.3",
-    "mini-css-extract-plugin": "^1.3.4",
     "normalizr": "^3.3.0",
     "postcss": "^8.2.4",
-    "postcss-loader": "^4.1.0",
     "rails-ujs": "5.2.4-5",
     "react": "^16.14.0",
     "react-beautiful-dnd": "^13.0.0",
@@ -33,22 +26,11 @@
     "react-toastify": "^7.0.3",
     "redux": "^4.0.5",
     "showdown": "^1.9.1",
-    "tailwindcss": "^2.0.2",
-    "webpack": "^5.11.0",
-    "webpack-cli": "^4.2.0"
-  },
-  "devDependencies": {
-    "@webpack-cli/serve": "^1.3.0",
-    "webpack-dev-server": "^3.11.2"
+    "tailwindcss": "^2.0.2"
   },
   "engines": {
-    "node": "14.16.0",
-    "yarn": "1.22.5"
-  },
-  "babel": {
-    "presets": [
-      "./node_modules/@rails/webpacker/package/babel/preset.js"
-    ]
+    "node": "18.14.0",
+    "yarn": "1.22.19"
   },
   "browserslist": [
     "defaults"
@@ -58,5 +40,10 @@
       "tailwindcss": {},
       "autoprefixer": {}
     }
+  },
+  "devDependencies": {
+    "@vitejs/plugin-react-swc": "^3.0.1",
+    "vite": "^4.0.0",
+    "vite-plugin-ruby": "^3.1.2"
   }
 }
