@@ -1,4 +1,4 @@
-import React, { useEffect, useState, useRef, useCallback } from "react";
+import { useEffect, useState, useRef, useCallback } from "react";
 import Octicon, { Clock } from "@githubprimer/octicons-react";
 import classNames from "classnames";
 
